More cleanups & refactoring
